+++ src/brain/include/brain_config.h
@@ -62,9 +62,12 @@ public:
     double headPitchLimitUp = 0.0;
 
     // Speed limit
-    double vxLimit = 1.2;
-    double vyLimit = 0.4;
-    double vthetaLimit = 1.5;
+    // double vxLimit = 1.2;
+    // double vyLimit = 0.4;    
+    double vxLimit = 2.0;
+    double vyLimit = 0.8;
+    // double vthetaLimit = 1.5;    
+    double vthetaLimit = 2.0;
 
     // Strategy parameters
     double safeDist = 2.0; // Safety distance for collision detection. If the distance is smaller than this value, a collision is considered.